backend_telegram/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
constantly==15.1.0
cryptography==40.0.2
dateparser==1.1.8
Django==4.2
django-appconf==1.0.5
django-widget-tweaks==1.4.12
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
gunicorn==20.0.4
msgpack==1.0.5
Pillow==9.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
python-dateutil==2.8.2
pytz==2023.3
psycopg2-binary==2.9.6
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
regex==2023.3.23
service-identity==21.1.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
zope.interface==6.0
djangorestframework==3.14.0
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.5
django-celery-beat==2.5.0
flower==2.0.1
django-filter==23.4